# contracts/cloud/openapi.yaml
openapi: 3.1.0
info:
  title: AO Cloud Public API
  version: 1.0.0
  description: |
    Client-facing contract for a separate AO Cloud implementation. This
    repository does not serve these routes. The worker surface documents the
    exact outbound HTTP protocol used by sandbox workers, including one-time
    bootstrap, rotating credentials, fenced turn execution, orchestration, and
    durable workspace and terminal transport.
security:
  - bearerAuth: []
...
paths:
...
  /api/cloud/v1/orgs/{orgId}/sessions:
    parameters:
      - $ref: "#/components/parameters/OrgId"
    get:
      operationId: listSessions
      tags: [Sessions]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: projectId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: A page of sessions in the tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionPage"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createSession
      tags: [Sessions]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionInput"
      responses:
        "201":
          description: Session created.
          content:
            application/json:
              schema:
                type: object
                required: [session]
                properties:
                  session:
                    $ref: "#/components/schemas/Session"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
    get:
      operationId: getSession
      tags: [Sessions]
      responses:
        "200":
          description: Session details.
          content:
            application/json:
              schema:
                type: object
                required: [session]
                properties:
                  session:
                    $ref: "#/components/schemas/Session"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteSession
      tags: [Sessions]
      responses:
        "202":
          description: Session sandbox deletion was requested; durable history is retained.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSessionResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    workerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Session-scoped credential formatted as `Worker <token>`. Every
        authenticated route verifies both the token signature and that its
        worker ID and epoch are still current. Replacing a sandbox therefore
        fences an otherwise unexpired token with `STALE_WORKER_TOKEN`.

        Route scopes are `worker:connect` for heartbeat, `worker:event` for
        events, `worker:turn:claim`, `worker:turn:poll`, and
        `worker:turn:complete` for turn execution,
        `worker:credential:read` for the agent secret, `worker:git` for a
        checkout grant, `worker:orchestrate` for child controls, and
        `worker:transport` for workspace and terminal transport, and
        `worker:notification` for durable agent notification intake.
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Reusing a key with the same command returns the original result.
        Reusing it with a different command returns an IDEMPOTENCY_CONFLICT.
      schema:
        type: string
        minLength: 1
        maxLength: 200
  responses:
    Error:
      description: Standard Cloud API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    AgentCapability:
      type: string
      enum:
        - interface.chat
        - interface.tui
        - model.catalog
        - model.custom
        - attachments
        - browser.preview
        - review.execute
        - session.resume
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, code, message, requestId]
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
    PageInfo:
      type: object
      additionalProperties: false
      required: [hasMore]
      properties:
        hasMore:
          type: boolean
        nextCursor:
          type: string
...
    SessionKind:
      type: string
      enum: [worker, orchestrator]
    SessionMode:
      type: string
      enum: [read-only, standard, trusted]
    SessionActivityState:
      type: string
      enum: [active, idle, waiting_input, blocked, exited]
    SessionStatus:
      type: string
      enum:
        - working
        - needs_input
        - pr_open
        - draft
        - review_pending
        - ci_failed
        - changes_requested
        - approved
        - mergeable
        - merged
        - exited
        - idle
        - terminated
        - no_signal
    Turn:
      type: object
      additionalProperties: false
      required: [id, sessionId, userMessageSequence, state, attemptCount, createdAt, updatedAt]
      properties:
        id:
          type: string
        sessionId:
          type: string
        userMessageSequence:
          type: integer
          format: int64
          maximum: 9007199254740991
        state:
          type: string
          enum: [queued, provisioning, running, cancel_requested, completed, failed]
        attemptCount:
          type: integer
          minimum: 0
        errorMessage:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    DeleteSessionResponse:
      type: object
      additionalProperties: false
      required: [session]
      properties:
        session:
          type: object
          additionalProperties: false
          required: [id, desiredState]
          properties:
            id:
              type: string
              format: uuid
            desiredState:
              type: string
              const: deleted
    Session:
      type: object
      additionalProperties: false
      required:
        - id
        - orgId
        - projectId
        - kind
        - harness
        - displayName
        - branch
        - mode
        - deniedCommands
        - activityState
        - status
        - runtimeConnected
        - isTerminated
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/SessionKind"
        harness:
          type: string
        displayName:
          type: string
        branch:
          type: string
        mode:
          $ref: "#/components/schemas/SessionMode"
        model:
          type: string
          description: >-
            Coding-agent model the session launched with; empty means the harness
            default.
        deniedCommands:
          type: array
          items:
            type: string
        activityState:
          $ref: "#/components/schemas/SessionActivityState"
        status:
          $ref: "#/components/schemas/SessionStatus"
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/AgentCapability"
        runtimeConnected:
          type: boolean
        runtimeState:
          type: string
        runtimeError:
          type: string
        activeTurn:
          $ref: "#/components/schemas/Turn"
        isTerminated:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSessionInput:
      type: object
      additionalProperties: false
      required: [projectId, kind, harness, displayName, prompt]
      properties:
        projectId:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/SessionKind"
        harness:
          type: string
          minLength: 1
          maxLength: 120
        displayName:
          type: string
          minLength: 1
          maxLength: 80
        prompt:
          type: string
          maxLength: 65536
        mode:
          $ref: "#/components/schemas/SessionMode"
          default: trusted
        model:
          type: string
          maxLength: 200
          description: >-
            Coding-agent model the session launches with (harness-native id, e.g.
            "anthropic/claude-opus-4-8"). Optional; empty uses the harness default.
        deniedCommands:
          type: array
          items:
            type: string
          default: []
        sandboxProviderConnectionId:
          type: string
          format: uuid
    SessionPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Session"
        page:
          $ref: "#/components/schemas/PageInfo"
...